precision fixes

+++ inav/flight/inav-2.0.0-RC4/src/main/io/gps.h
@@ -43,7 +43,7 @@ typedef struct gpsDataPozyx_s {
     int y_sign;
     int z_sign;
 
-    int32_t altitude;
+    float altitude;
 } gpsDataPozyx_t;
 
 typedef enum {
@@ -120,9 +120,9 @@ typedef struct gpsCoordinateDDDMMmmmm_s {
 
 /* LLH Location in NEU axis system */
 typedef struct gpsLocation_s {
-    int32_t lat;    // Latitude * 1e+7
-    int32_t lon;    // Longitude * 1e+7
-    int32_t alt;    // Altitude in centimeters (meters * 100)
+    int32_t lat;    // Latitude * 1e+6
+    int32_t lon;    // Longitude * 1e+6
+    float alt;    // Altitude in meters
     int32_t x;
     int32_t y;
     int32_t z;
